Original push
This push features all notebooks (named Figure1-8) which contains the code used to generate all analyses and figures used in the manuscript.

Intermediate data structures are stored in the folder called "intermediates" and sometimes accessed by the code to skip blocks that may take time to generate.

The recordings folder contains data pertinent to each recording that are used. Each recording has its own folder which is accessed in the code by mouse name and recording date. This github repository does not include the NWBs which can be downloaded on DANDI. To run this code independently, please download the NWBs and place them in the corresponding recording folder.

Lastly, there's some broad wrapper and analysis tools included in the folder "jlh_ephys" which can be imported when run locally which faciliated some analysis and loading of data.